Separation of static API starting point and mapping builder
