+++ src/main/java/com/remondis/remap/OmitTransformation.java
@@ -14,7 +14,7 @@ class OmitTransformation extends Transformation {
 
   private static final String OMITTING_MSG = "Omitting %s";
 
-  private OmitTransformation(Mapping<?, ?> mapping, PropertyDescriptor sourceProperty,
+  private OmitTransformation(MappingBuilder<?, ?> mapping, PropertyDescriptor sourceProperty,
       PropertyDescriptor destinationProperty) {
     super(mapping, sourceProperty, destinationProperty);
   }
@@ -26,7 +26,7 @@ class OmitTransformation extends Transformation {
    * @param destinationProperty the property to omit in the destination
    * @return Returns a new {@link OmitTransformation}.
    */
-  static OmitTransformation omitDestination(Mapping<?, ?> mapping, PropertyDescriptor destinationProperty) {
+  static OmitTransformation omitDestination(MappingBuilder<?, ?> mapping, PropertyDescriptor destinationProperty) {
     denyNull("mapping", mapping);
     denyNull("destinationProperty", destinationProperty);
     return new OmitTransformation(mapping, null, destinationProperty);
@@ -39,7 +39,7 @@ class OmitTransformation extends Transformation {
    * @param sourceProperty the property to omit in the source
    * @return Returns a new {@link OmitTransformation}.
    */
-  static OmitTransformation omitSource(Mapping<?, ?> mapping, PropertyDescriptor sourceProperty) {
+  static OmitTransformation omitSource(MappingBuilder<?, ?> mapping, PropertyDescriptor sourceProperty) {
     denyNull("mapping", mapping);
     denyNull("destinationProperty", sourceProperty);
     return new OmitTransformation(mapping, sourceProperty, null);
