+++ src/main/java/com/remondis/remap/ReplaceTransformation.java
@@ -22,7 +22,7 @@ class ReplaceTransformation<RS, RD> extends SkipWhenNullTransformation<RS, RD> {
   private Function<RS, RD> transformation;
   private boolean skipWhenNull;
 
-  ReplaceTransformation(Mapping<?, ?> mapping, PropertyDescriptor sourceProperty, PropertyDescriptor destProperty,
+  ReplaceTransformation(MappingBuilder<?, ?> mapping, PropertyDescriptor sourceProperty, PropertyDescriptor destProperty,
       Function<RS, RD> transformation, boolean skipWhenNull) {
     super(mapping, sourceProperty, destProperty);
     this.transformation = transformation;
