+++ src/main/java/com/remondis/remap/Transformation.java
@@ -17,9 +17,9 @@ abstract class Transformation {
 
   protected PropertyDescriptor sourceProperty;
   protected PropertyDescriptor destinationProperty;
-  private Mapping<?, ?> mapping;
+  private MappingBuilder<?, ?> mapping;
 
-  Transformation(Mapping<?, ?> mapping, PropertyDescriptor sourceProperty, PropertyDescriptor destinationProperty) {
+  Transformation(MappingBuilder<?, ?> mapping, PropertyDescriptor sourceProperty, PropertyDescriptor destinationProperty) {
     super();
     denyNull("mapping", mapping);
     this.mapping = mapping;
